# [Simulator] Randomly permute gates before the simple DP